PWN-3247 - add TransactionSwapImage.kt and use it in TransactionSwapViewHolder.kt
