+++ app/src/main/res/layout/widget_transaction_image.xml
@@ -15,7 +15,7 @@
         android:padding="12dp"
         android:layout_margin="2dp"
         android:src="@drawable/ic_transaction_send"
-        app:cornerRadius="12dp"
+        app:cornerRadius="16dp"
         tools:visibility="visible" />
 
     <ImageView
