+++ app/src/main/res/layout/widget_transaction_swap_image.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<merge
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:parentTag="android.widget.FrameLayout">
+
+    <View
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_margin="2dp"
+        android:background="@drawable/bg_token_swap_rounded" />
+
+    <org.p2p.wallet.common.ui.widget.RoundedImageView
+        android:id="@+id/sourceImageView"
+        android:layout_width="29dp"
+        android:layout_height="29dp"
+        android:layout_gravity="top"
+        app:cornerRadius="9dp"
+        tools:src="@tools:sample/avatars" />
+
+    <org.p2p.wallet.common.ui.widget.RoundedImageView
+        android:id="@+id/destinationImageView"
+        android:layout_width="29dp"
+        android:layout_height="29dp"
+        android:layout_gravity="end|bottom"
+        app:cornerRadius="9dp"
+        tools:src="@tools:sample/avatars" />
+
+</merge>
